datapath/linux/ethtool: use IoctlGetEthtoolDrvinfo
[ upstream commit d82ac6f54c0118088cc46d8d892ff5e87cf5d09e ]

Use the ioctl wrapper provided in the golang.org/x/sys/unix package with
the correctly padded ifreqData struct, rather than providing our own
wrapper and struct which is incorrectly padded.

Also add a simple unit test and make sure the package is only built on
Linux.
